refactor(admin): move themes store to runes
